README.md tweaks: C++11 initialision, Markdown.
Suggesting a few cosmetic changes to the README.  The changes require
C++11 as the minimum, but I hope this won't be a problem since the
library depends on libpqxx which currently requires C++17 or better.

Most of the tweaks are about the Markdown syntax:
* Use C++11 initialisation syntax for those `std::map`s.  Much nicer.
* Mark code blocks as being in specific languages, where the choice is
  clear.  May enable syntax highlighting, depending on the renderer.
* Remove trailing whitespace.
* Escape underscores that are not meant to denote italics.

I found most of the Markdown gotchas by running `mdl`, a Markdown lint
checker.